move ACPick from discard stuff to ACPickFromDiscard
