+++ src/main/java/ti4/commands/cardsac/PickACFromDiscard.java
@@ -1,15 +1,22 @@
 package ti4.commands.cardsac;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
 import net.dv8tion.jda.api.events.interaction.GenericInteractionCreateEvent;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
 import net.dv8tion.jda.api.interactions.commands.OptionMapping;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.OptionData;
+import net.dv8tion.jda.api.interactions.components.buttons.Button;
+import ti4.buttons.Buttons;
 import ti4.generator.Mapper;
+import ti4.helpers.ButtonHelper;
 import ti4.helpers.Constants;
 import ti4.helpers.Helper;
+import ti4.listeners.annotations.ButtonHandler;
 import ti4.map.Game;
 import ti4.map.Player;
 import ti4.message.MessageHelper;
@@ -63,5 +70,43 @@ public class PickACFromDiscard extends ACCardsSubcommandData {
         MessageHelper.sendMessageToChannel(event.getMessageChannel(), sb);
 
         ACInfo.sendActionCardInfo(game, player);
+    }
+
+    public static void pickACardFromDiscardStep1(Game game, Player player) {
+        List<Button> buttons = new ArrayList<>();
+        for (String acStringID : game.getDiscardActionCards().keySet()) {
+            buttons.add(Buttons.green("pickFromDiscard_" + acStringID, Mapper.getActionCard(acStringID).getName()));
+        }
+        buttons.add(Buttons.red("deleteButtons", "Delete These Buttons"));
+        if (buttons.size() > 25) {
+            buttons.add(25, Buttons.red("deleteButtons_", "Delete These Buttons"));
+        }
+        if (buttons.size() > 50) {
+            buttons.add(50, Buttons.red("deleteButtons_2", "Delete These Buttons"));
+        }
+        if (buttons.size() > 75) {
+            buttons.add(75, Buttons.red("deleteButtons_3", "Delete These Buttons"));
+        }
+        String msg = player.getRepresentation(true, true) + " use buttons to grab an AC from the discard";
+        MessageHelper.sendMessageToChannelWithButtons(player.getCardsInfoThread(), msg, buttons);
+    }
+
+    @ButtonHandler("pickFromDiscard_")
+    public static void pickACardFromDiscardStep2(Game game, Player player, ButtonInteractionEvent event,        String buttonID) {
+        ButtonHelper.deleteMessage(event);
+        String acID = buttonID.replace("pickFromDiscard_", "");
+        boolean picked = game.pickActionCard(player.getUserID(), game.getDiscardActionCards().get(acID));
+        if (!picked) {
+            MessageHelper.sendMessageToChannel(event.getChannel(), "No such Action Card ID found, please retry");
+            return;
+        }
+        String msg2 = player.getRepresentation(true, true) + " grabbed " + Mapper.getActionCard(acID).getName() + " from the discard";
+        MessageHelper.sendMessageToChannel(player.getCorrectChannel(), msg2);
+
+        ACInfo.sendActionCardInfo(game, player, event);
+        if (player.hasAbility("autonetic_memory")) {
+            String message = player.getRepresentation(true, true) + " if you did not just use the Codex to get that AC, please discard 1 AC due to your Cybernetic Madness ability";
+            MessageHelper.sendMessageToChannelWithButtons(player.getCardsInfoThread(), message, ACInfo.getDiscardActionCardButtons(game, player, false));
+        }
     }
 }
